+++ mongo/gridfs/download_stream.go
@@ -260,7 +260,7 @@ func (ds *DownloadStream) fillBuffer(ctx context.Context) error {
 	if ds.expectedChunk == ds.numChunks {
 		// final chunk can be fewer than ds.chunkSize bytes
 		bytesDownloaded := int64(ds.chunkSize) * (int64(ds.expectedChunk) - int64(1))
-		bytesRemaining := ds.fileLen - int64(bytesDownloaded)
+		bytesRemaining := ds.fileLen - bytesDownloaded
 
 		if int64(bytesLen) != bytesRemaining {
 			return ErrWrongSize
